Split Linux GPU renderer into Mesa and non-Mesa pipelines
